Changes from Aaron
Switched to GRanges-like formatting, rather than reporting as a compressed string. The idea is to ensure chromosome names and coordinates are aligned between entries, whereas putting everything in a single string makes it a bit unpretty.
Switched to separate 'anchors' calls; for some reason, calling 'anchors(x)' results in a delay, probably due to GRangesList construction.
Replaced all "=" with "<-" for assignment.
Removed the extraneous one-space offset in the seqinfo separator.